+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added csrf protection across all routes. For custom forms, the user will need to add `csrfToken: req.csrfToken()` to the options when using `linz.api.model.generateFormString()`.
 - Csrf protection can be customised by setting the linz option `csrf options`. This takes an object with the same options listed here https://www.npmjs.com/package/csurf.
 - Fixed not being able to return html in an overview section.
+- Export fields are no longer required to be in the formDSL.
 
 ## v1.0.0-17.1.1